feat:add new idea form modal

+++ src/controllers/idea.controller.ts
@@ -1,43 +1,48 @@
 import type { DestroyableController } from '../interfaces/destroyable-controller.interface'
-import { editIdeaForm } from '../render/edit-idea-form'
+import { editIdeaFormRender } from '../render/edit-idea-form.render'
 import { ideaCardRender } from '../render/idea-card.render'
 
 import type { IdeaService } from '../services/idea.service'
 import type { ModalService } from '../services/modal.service'
 import { UpdateIdeaSchema } from '../schemas/idea/update-idea.schema'
+import { CreateIdeaSchema } from '../schemas/idea/create-idea.schema'
+import { newIdeaFormRender } from '../render/new-idea-form.render'
+import { render } from '../render/render'
 
 export class IdeaController implements DestroyableController {
   private modalService: ModalService
   private service: IdeaService
   private container: HTMLElement
-  private actions: Record<string, (id: string) => Promise<void> | void>
-  /*   private boundHandlerClick: (ev: PointerEvent) => Promise<void>
-  private boundHandlerSubmit: (ev: SubmitEvent) => Promise<void> */
+  private actions: Record<string, (id?: string) => Promise<void> | void>
 
   constructor(service: IdeaService, container: HTMLElement, modalService: ModalService) {
     this.modalService = modalService
     this.service = service
     this.container = container
-    //binds  X  refactor arrow f
-    /* this.boundHandlerClick = this.handlerClick.bind(this)
-    this.boundHandlerSubmit = this.handlerSubmit.bind(this) */
 
     //actions
     this.actions = {
       edit: this.editIdea,
       delete: this.deleteIdea,
+      new: this.newIdea,
     }
-    console.log(this.actions)
-    this.setupDelegation()
+
+    this.container.addEventListener('click', this.handlerClick) //setup delegation click
     this.listIdea()
   }
 
   handlerClick = async (ev: PointerEvent) => {
     ev.preventDefault()
     const target = ev.target as HTMLElement
-
     const action = target.dataset.action
     if (!action) {
+      return
+    }
+
+    //new not need to id
+    const handler = this.actions[action]
+    if (action === 'new') {
+      await handler()
       return
     }
 
@@ -45,32 +50,22 @@ export class IdeaController implements DestroyableController {
     if (!card) {
       return
     }
-    const id = card.getAttribute('data-id')
+    const id = card.getAttribute('data-id') || undefined
     if (!id) {
       return
     }
     console.log('actions and action', this.actions, action)
-    const handler = this.actions[action]
+
     if (handler) {
       await handler(id)
     }
   }
 
-  private handlerSubmit = async (ev: SubmitEvent) => {
-    ev.preventDefault()
-
-    const target = ev.target as HTMLFormElement
-    const formData = new FormData(target)
-    console.log(formData)
-  }
-  setupDelegation() {
-    console.log('setup delegation idea')
-    this.container.addEventListener('submit', this.handlerSubmit)
-    //click
-    this.container.addEventListener('click', this.handlerClick)
-  }
-
-  private deleteIdea = async (id: string): Promise<void> => {
+  //actions click
+  private deleteIdea = async (id?: string): Promise<void> => {
+    if (!id) {
+      return
+    }
     console.log('delete idea')
     await this.service.destroy(id)
     const card = document.getElementById(id)
@@ -80,12 +75,37 @@ export class IdeaController implements DestroyableController {
     }
   }
 
-  private editIdea = async (id: string) => {
-    const exit = await this.service.getById(id)
-    if (exit) {
-      const form = editIdeaForm(exit)
+  private newIdea = async () => {
+    const form = newIdeaFormRender()
+    //register handler y new form show modal
+    this.modalService.registerHandler('click', (ev: Event) => {
+      if (ev as PointerEvent) {
+        this.modalService.closeModal()
+      }
+    })
+    this.modalService.registerHandler('submit', (ev: Event) =>
+      this.handleSubmitNewModal(ev as SubmitEvent)
+    )
+    this.modalService.showModal(form)
+  }
 
-      this.modalService.registerHandler('submit', this.handlerSubmitEditModal)
+  private editIdea = async (id?: string) => {
+    if (!id) {
+      return
+    }
+    const exit = await this.service.getById(id)
+
+    if (exit) {
+      const form = editIdeaFormRender(exit)
+
+      this.modalService.registerHandler('click', (ev: Event) => {
+        if (ev as PointerEvent) {
+          this.modalService.closeModal()
+        }
+      })
+      this.modalService.registerHandler('submit', (ev: Event) =>
+        this.handlerSubmitEditModal(ev as SubmitEvent)
+      )
       /* form.addEventListener('submit', this.handlerSubmitEdit,{once:true}) */
       this.modalService.showModal(form)
 
@@ -93,41 +113,88 @@ export class IdeaController implements DestroyableController {
     }
   }
 
-  private handlerSubmitEditModal = async (ev: Event) => {
+  // handler form modal
+  private handleSubmitNewModal = async (ev: SubmitEvent) => {
     ev.preventDefault()
 
     const target = ev.target as HTMLFormElement
     const formData = new FormData(target)
-    console.log(formData)
+
+    const title = formData.get('title')?.toString()
+    const description = formData.get('description')?.toString()
+
+    console.log('new idea with', title, description)
+    const schema = CreateIdeaSchema.safeParse({ title, description })
+    if (schema.success) {
+      try {
+        const newIdea = await this.service.create(schema.data)
+
+        //card append
+        const cardNew = ideaCardRender(newIdea)
+        console.log('card new ', cardNew)
+        const div = this.container.querySelector('#idea-list-div')
+        div?.prepend(cardNew)
+      } catch (error) {
+        console.log('Error new idea input', error)
+      }
+    } else {
+      alert('data invalid')
+    }
+
+    this.modalService.closeModal()
+  }
+  private handlerSubmitEditModal = async (ev: SubmitEvent) => {
+    ev.preventDefault()
+
+    const target = ev.target as HTMLFormElement
+    const formData = new FormData(target)
+
     const title = formData.get('title')?.toString()
     const description = formData.get('description')?.toString()
     const id = formData.get('id')?.toString()
     const created = formData.get('created')?.toString()
-    //todo edit service call
+
     console.log('edit idea with', title, description, id, created)
-    //delete handler ev.target.remove->handlerSubmitEdit?
+    const card = document.getElementById(id || '')
+    if (!card) {
+      console.log('no exist card id', id)
+
+      return
+    }
+
     const update = UpdateIdeaSchema.safeParse({ _id: id, title, description })
+
     if (update.success) {
-      console.log(await this.service.update(update.data))
-      //todo buscar card y remplazar con nueva data
+      try {
+        const ideaEdit = await this.service.update(update.data)
+        //remplace card
+        const cardEdit = ideaCardRender(ideaEdit)
+        console.log('edit card', cardEdit)
+        card.replaceWith(cardEdit)
+      } catch (error) {
+        console.log('Error edit idea input', error)
+      }
+    } else {
+      alert('data invalid')
     }
 
     this.modalService.closeModal()
   }
 
+  //init contruct
   private listIdea = async () => {
     const list = await this.service.getAll()
     console.log(list)
+    const div = this.container.querySelector('#idea-list-div') as HTMLElement
+    if (div) {
+      const frag = document.createDocumentFragment()
+      list.forEach(l => frag.append(ideaCardRender(l)))
 
-    const div = this.container.querySelector('#idea-list-div')
-    console.log(div)
-    list.forEach(l => {
-      div?.append(ideaCardRender(l))
-    })
+      render(div, frag)
+    }
   }
   //interface
   destroy(): void {
-    this.container.removeEventListener('submit', this.handlerSubmit)
     this.container.removeEventListener('click', this.handlerClick)
   }
 }
